+++ src/extension.ts
@@ -1,8 +1,5 @@
 import * as path from 'path';
-import {
-  // workspace,
-  ExtensionContext,
-} from 'vscode';
+import * as vscode from 'vscode';
 import {
   LanguageClient,
   LanguageClientOptions,
@@ -10,9 +7,7 @@ import {
   TransportKind,
 } from 'vscode-languageclient/node';
 
-let client: LanguageClient;
-
-export function activate(context: ExtensionContext): void {
+export function activate(context: vscode.ExtensionContext): void {
   const serverModule = context.asAbsolutePath(
     path.join(
       'node_modules',
@@ -22,6 +17,36 @@ export function activate(context: ExtensionContext): void {
       'server.js'
     )
   );
+
+  const configuredView = vscode.workspace.getConfiguration();
+
+  context.subscriptions.push(vscode.commands.registerCommand('superfaceLanguageClient.commands.languageServer.restart', () => {
+    runClient(serverModule, true);
+  }));
+
+  context.subscriptions.push(vscode.commands.registerCommand('superfaceLanguageClient.commands.languageServer.stop', async () => {
+    return stopClient();
+  }));
+
+  if (configuredView.get('superfaceLanguageClient.languageServer.enabled') === true) {
+    runClient(serverModule);
+  }
+}
+
+export function deactivate(): Promise<void> | undefined {
+  return stopClient();
+}
+
+
+let CLIENT: LanguageClient | undefined;
+function runClient(serverModule: string, force?: boolean): boolean {
+  if (CLIENT !== undefined) {
+    if (force === true) {
+      stopClient();
+    } else {
+      return false;
+    }
+  }
 
   // If the extension is launched in debug mode then the debug server options are used
   // Otherwise the run options are used
@@ -52,21 +77,25 @@ export function activate(context: ExtensionContext): void {
   };
 
   // Create the language client and start the client.
-  client = new LanguageClient(
+  CLIENT = new LanguageClient(
     'superfaceLanguageClient',
     'Superface Language Client',
     serverOptions,
     clientOptions
   );
 
-  // Start the client. This will also launch the server
-  client.start();
+  // Start the client. This will also launch the server.
+  CLIENT.start();
+
+  return true;
 }
 
-export function deactivate(): Promise<void> | undefined {
-  if (!client) {
+function stopClient(): undefined | Promise<void> {
+  if (CLIENT === undefined) {
     return undefined;
   }
 
-  return client.stop();
+  const result = CLIENT.stop();
+  CLIENT = undefined;
+  return result;
 }